Fix compatibility with mod_rewrite in proxy mode
mod_rewrite passes cookies to the backend servers when run in proxy
mode. For us this is helpful as our single-sign-on solution is cookie
based. We use mod_rewrite to map backend APIs to our frontend servers
and thus fix the CORS problem.

When testing urlrewritefilter as an alternative solution I noticed,
that my requests were all redirected to our authentication servers and
the dropped cookies would cause that behavior. This triggered my
investigation.